refactor: moved type definitions to its own file
